requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.10.4
boto==2.38.0
gunicorn==19.3.0
httplib2==0.9.1
itsdangerous==0.24
parse==1.6.6
parse-rest==0.2.20141004
pytz==2015.4
six==1.9.0
twilio==4.5.0
wsgiref==0.1.2
python-postmark>=0.4.3
